refactor(control): split nfqueue ingest
